fix: fixed some tests by ensuring the dir exits before removing

+++ databox/core/test/Databox.queryInternal.test.ts
@@ -22,7 +22,7 @@ beforeAll(async () => {
 });
 
 afterAll(async () => {
-  Fs.rmdirSync(storageDir, { recursive: true });
+  if (Fs.existsSync(storageDir)) Fs.rmdirSync(storageDir, { recursive: true });
   await miner.close();
 });
